[codemirror] extracted mode services to mode module
